feat(src/App.jsx and components/*): add Section and Button modules
